put all course related pages in courses folder
