+++ App/src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Navigate, Route, Routes, useNavigate } from "react-route
 
 // 1. import `ChakraProvider` component
 import { ChakraProvider } from "@chakra-ui/react";
-import CourseBrowser from "./pages/CourseBrowser";
+import CourseBrowser from "./pages/courses/CourseBrowser";
 import Home from "./pages/Home";
 import NotFound from "./pages/NotFound";
 import Login from "./pages/Login";
@@ -10,7 +10,7 @@ import { User } from "./models/user";
 import { createContext, useState } from "react";
 import Register from "./pages/Register";
 import { decodeTokenToUser, isAdmin, isAdminOrGroupMaster } from "./services/auth";
-import AddCourse from "./pages/AddCourse";
+import AddCourse from "./pages/courses/AddCourse";
 import theme from "../theme"; // Import your custom theme
 import ComponentsList from "./pages/ComponentsList";
 import { LoadingProvider } from "./utils/loadingContext"; // Import the provider
@@ -20,14 +20,14 @@ import Dossiers from "./pages/dossier/Dossiers";
 import DisplayManageableGroups from "./pages/groups/ManageableGroups";
 import AddingUserToGroup from "./pages/groups/addUserToGroup";
 import RemovingUserFromGroup from "./pages/groups/RemoveUserFromGroup";
-import DeleteCourse from "./pages/DeleteCourse";
+import DeleteCourse from "./pages/courses/DeleteCourse";
 import DossierDetails from "./pages/dossier/DossierDetails";
 import AddingMasterToGroup from "./pages/groups/AddGroupMaster";
 import RemovingMasterFromGroup from "./pages/groups/RemoveGroupMaster";
 import DossiersToReview from "./pages/dossier/DossierToReview";
 import CreateGroup from "./pages/groups/CreateGroup";
-import DeleteCourseEdit from "./pages/DeleteCourseEdit";
-import CourseDetails from "./pages/CourseDetails";
+import DeleteCourseEdit from "./pages/courses/DeleteCourseEdit";
+import CourseDetails from "./pages/courses/CourseDetails";
 import Header from "./shared/Header";
 import DossierReview from "./pages/dossier/DossierReview";
 import DossierReport from "./pages/dossier/DossierReport";
@@ -50,7 +50,7 @@ import CoursesLeft from "./pages/CoursesLeft";
 import UserBrowser from "./pages/UserBrowser";
 import EmailResetPassword from "./pages/EmailResetPassword";
 import Directories from "./pages/Directories";
-import CourseBySubject from "./pages/CourseBySubjectBrowser";
+import CourseBySubject from "./pages/courses/CourseBySubjectBrowser";
 import CoursesFromSubject from "./pages/CoursesFromSubject";
 
 export const UserContext = createContext<User | null>(null);
